# src/docs/swagger.yaml
basePath: /
definitions:
  model.CostResponse:
    properties:
      total_cost:
        type: integer
    type: object
  model.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 0
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  model.Subscription:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      price:
        minimum: 0
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  model.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 0
        type: integer
      service_name:
        type: string
      start_date:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST-сервис для агрегации данных об онлайн подписках
  title: Subscription Service API
  version: "1.0"
paths:
  /api/v1/subscriptions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
      summary: List all subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      parameters:
      - description: Subscription
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create subscription
      tags:
      - subscriptions
  /api/v1/subscriptions/{id}:
    delete:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update subscription
      tags:
      - subscriptions
  /api/v1/subscriptions/cost:
    get:
      parameters:
      - description: Start period (MM-YYYY)
        in: query
        name: from
        required: true
        type: string
      - description: End period (MM-YYYY)
        in: query
        name: to
        required: true
        type: string
      - description: User ID (UUID)
        in: query
        name: user_id
        type: string
      - description: Service name
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CostResponse'
      summary: Get total cost of subscriptions
      tags:
      - subscriptions
swagger: "2.0"